Add method for getting a list of recordable iOS screens
Not included in the JS code yet, as we still need to add support for recording the iOS screen. It's not considered a screen, but rather a normal video input, so we need to add some custom logic for that.

Also clean up the code for getting audio devices.

Part of wulkano/kap#71